fix(metrics): wire missing fields and add comprehensive tests
FrameworkMetrics::snapshot() returned hardcoded zeros for 6 fields.
This was due to FrameworkMetrics not having access to the internal
counters of Singularity (cache) and Reservoir components.

Refactored metrics to use shared ownership via Arc:
- FrameworkMetrics now owns Arc<CacheMetrics> and Arc<ReservoirMetrics>.
- Singularity and Reservoir are initialized with these shared instances.
- metrics_snapshot() now returns live data for all 12 fields.
- Added persistence metrics (ops count and avg latency) to snapshot.
- Updated WASM bindings to include all 12 snapshot fields.

Resolved CI failures and improved robustness:
- Moved ChaoticReservoir to src/reservoir_chaotic.rs to satisfy 500 LOC gate.
- Fixed clippy::float-cmp by using epsilon comparisons in tests.
- Fixed clippy::clone-on-copy in tests.
- Applied cargo fmt workspace-wide.

Added comprehensive integration tests:
- Verified multi-namespace cache metrics aggregation.
- Verified reservoir metrics wiring and live updates.
- Verified zero-division safety for latency averages.

Verified with full test suite and clippy.
