+++ .github/workflows/ci.yml
@@ -125,10 +125,6 @@ jobs:
       uses: Swatinem/rust-cache@c19371144df3bb44fab255c43d04cbc2ab54d1c4 # v2
     - name: Check WASM
       run: cargo check --target wasm32-unknown-unknown --features wasm
-    - name: Install wasm-pack
-      run: curl https://rustwasm.github.io/wasm-pack/installer/init.sh -sSf | sh
-    - name: Verify WASM Freshness
-      run: bash scripts/check-wasm-freshness.sh
 
   lint:
     runs-on: ubuntu-latest
