--- export.json
@@ -1,6 +1,0 @@
-{
-  "version": "0.3.6",
-  "exported_at": 1779642716,
-  "concepts": [],
-  "associations": []
-}
