+++ src/wasm_graph_rag.rs
@@ -8,15 +8,12 @@ use js_sys::Array;
 use wasm_bindgen::prelude::*;
 
 #[cfg(target_arch = "wasm32")]
-use crate::wasm::WasmFramework;
-#[cfg(target_arch = "wasm32")]
-use crate::wasm_utils::to_js_error;
+use crate::wasm::{WasmFramework, to_js_error};
 
 #[cfg(target_arch = "wasm32")]
 #[wasm_bindgen]
 impl WasmFramework {
     /// GraphRAG retrieval: similarity + graph traversal hybrid using vector query.
-    #[wasm_bindgen(js_name = probe_with_graph)]
     pub async fn probe_with_graph(
         &self,
         vector: &[u8],
@@ -68,7 +65,6 @@ impl WasmFramework {
     }
 
     /// GraphRAG retrieval: similarity + graph traversal hybrid using text query.
-    #[wasm_bindgen(js_name = probe_text_with_graph)]
     pub async fn probe_text_with_graph(
         &self,
         text: String,
@@ -122,7 +118,6 @@ impl WasmFramework {
     ///
     /// Returns an Array of `{id: string, depth: number}` objects.
     /// Uses default `TraversalConfig`.
-    #[wasm_bindgen(js_name = bfs)]
     pub async fn bfs(&self, start: String) -> Result<Array, JsValue> {
         use crate::graph_traversal::TraversalConfig;
         let sing = self.framework.singularity.read().await;
@@ -145,7 +140,6 @@ impl WasmFramework {
     ///
     /// Returns an Array of concept ID strings, or an empty Array if no path exists.
     /// Uses default `TraversalConfig`.
-    #[wasm_bindgen(js_name = shortest_path)]
     pub async fn shortest_path(&self, from: String, to: String) -> Result<Array, JsValue> {
         let path = self
             .framework
@@ -162,7 +156,6 @@ impl WasmFramework {
     }
 
     /// Breadth-first traversal from a starting concept with custom config.
-    #[wasm_bindgen(js_name = traverse)]
     pub async fn traverse(
         &self,
         start: String,
